lawful-classes-{hedgehog,quickcheck}: add `testLawsWith`
The new `testLawsWith` functions allow to modify properties (as created
by `testLaws`) before they're added to the `TestTree`. As an example,
they allow to run tests only once by passing `Hedgehog.withTests 1` or
`Test.QuickCheck.once` as a modifier. The latter is useful when testing
laws against `Monad`s that are no transformers, and hence can't generate
exemplars, but must be seeded with some constant test values (as now
demonstrated in `lawful-classes-demo`).

Given this change, the versions of both libraries are bumped to 0.1.1.0.