lawful-classes-demo: require lawful-classes-{hedgehog,quickcheck} ^>=0.1.1.0
Since the demo code is using `testLawsWith` since 799bed726ebc0, the
tests rely on versions of lawful-classes-{hedgehog,quickcheck} which
expose this function. This was not reflected in the `build-depends` of
`lawful-classes-demo-test` due to an oversight, and not caught by CI
since builds occur against the in-tree versions of said libraries.

See: 799bed726ebc0b5cabd3652bbc774e38c8a8ecc5